change: Move DAO methods on StatusViewData in to StatusDao
It's a better fit for them than TimelineDao.
